Rename package messages.Messages to message.Message
