+++ src/main/java/baron/commands/AddTaskCommand.java
@@ -1,7 +1,7 @@
 package baron.commands;
 
 import baron.exceptions.BaronException;
-import baron.messages.Messages;
+import baron.message.Message;
 import baron.tasks.Task;
 import baron.tasks.TaskManager;
 import baron.tasks.TaskType;
@@ -22,7 +22,7 @@ public class AddTaskCommand extends Command {
 
     public String execute() {
         if (this.commandArg.isBlank()) {
-            return (new BaronException(Messages.generateEmptyDescMessage(this.taskType))).toString();
+            return (new BaronException(Message.generateEmptyDescMessage(this.taskType))).toString();
         }
 
         try {
@@ -34,8 +34,8 @@ public class AddTaskCommand extends Command {
                 throw e;
             }
             this.taskManager.commitChanges();
-            return Messages.MESSAGE_ADD_TASK_SUCCESS + addedTask.toString() + "\n"
-                    + Messages.generateNoOfTasksMessage(this.taskManager.getTaskCount());
+            return Message.MESSAGE_ADD_TASK_SUCCESS + addedTask.toString() + "\n"
+                    + Message.generateNoOfTasksMessage(this.taskManager.getTaskCount());
         } catch (BaronException e) {
             return e.toString();
         }
